+++ src/EnrolmentWorkflow/Action/AssignResponsibleUsers/AssignResponsibleUsersRunner.php
@@ -88,7 +88,7 @@ class AssignResponsibleUsersRunner extends AbstractActionRunner
         }
 
         $responsible_users = array_filter($responsible_users, function (int $user_id) use ($request): bool {
-            return ($request->getUserId() !== $user_id && count(self::srUserEnrolment()->enrolmentWorkflow()->requests()->getRequests($request->getObjRefId(), null, $user_id)) < 2);
+            return ($request->getUserId() !== $user_id && count(self::srUserEnrolment()->enrolmentWorkflow()->requests()->getRequests(null, $request->getObjRefId(), null, $user_id)) < 2);
         });
 
         $request->setResponsibleUsers($responsible_users);