+++ src/EnrolmentWorkflow/Request/AllRequestsTableGUI.php
@@ -112,6 +112,15 @@ class AllRequestsTableGUI extends AbstractRequestsTableGUI
     /**
      * @inheritDoc
      */
+    protected function getFilterCheckUserId()/* : ?int*/
+    {
+        return self::dic()->user()->getId();
+    }
+
+
+    /**
+     * @inheritDoc
+     */
     protected function getFilterObjRefId()/* : ?int*/
     {
         return $this->parent_obj->getObjRefId();
@@ -156,7 +165,7 @@ class AllRequestsTableGUI extends AbstractRequestsTableGUI
     /**
      * @inheritDoc
      */
-    protected function getFilterUsrId()/* : ?int*/
+    protected function getFilterUserId()/* : ?int*/
     {
         return null;
     }