+++ src/EnrolmentWorkflow/Request/OpenRequestsTableGUI.php
@@ -89,6 +89,15 @@ class OpenRequestsTableGUI extends AbstractRequestsTableGUI
     /**
      * @inheritDoc
      */
+    protected function getFilterCheckUserId()/* : ?int*/
+    {
+        return self::dic()->user()->getId();
+    }
+
+
+    /**
+     * @inheritDoc
+     */
     protected function getFilterObjRefId()/* : ?int*/
     {
         return $this->parent_obj->getObjRefId();
@@ -129,7 +138,7 @@ class OpenRequestsTableGUI extends AbstractRequestsTableGUI
     /**
      * @inheritDoc
      */
-    protected function getFilterUsrId()/* : ?int*/
+    protected function getFilterUserId()/* : ?int*/
     {
         return null;
     }