+++ src/EnrolmentWorkflow/Request/class.RequestsGUI.php
@@ -165,14 +165,17 @@ class RequestsGUI
             self::dic()->toolbar()->addButtonInstance($request_button);
         }
 
-        self::dic()->ctrl()->setParameterByClass(self::class, self::GET_PARAM_REQUESTS_TYPE, self::REQUESTS_TYPE_ALL);
-        self::dic()->tabs()->addTab(self::TAB_REQUESTS . self::REQUESTS_TYPE_ALL, self::plugin()->translate("requests", self::LANG_MODULE), self::dic()->ctrl()
-            ->getLinkTargetByClass([ilUIPluginRouterGUI::class, self::class], self::CMD_LIST_REQUESTS));
+        if (!empty($this->obj_ref_id)) {
+            self::dic()->ctrl()->setParameterByClass(self::class, self::GET_PARAM_REQUESTS_TYPE, self::REQUESTS_TYPE_ALL);
+            self::dic()->tabs()->addTab(self::TAB_REQUESTS . self::REQUESTS_TYPE_ALL, self::plugin()->translate("requests", self::LANG_MODULE), self::dic()->ctrl()
+                ->getLinkTargetByClass([ilUIPluginRouterGUI::class, self::class], self::CMD_LIST_REQUESTS));
+        }
 
         foreach (self::REQUESTS_TYPES as $requests_type => $requests_type_lang_key) {
             self::dic()->ctrl()->setParameter($this, self::GET_PARAM_REQUESTS_TYPE, $requests_type);
-            self::dic()->tabs()->addSubTab(self::TAB_REQUESTS . $requests_type, self::plugin()->translate("type_" . $requests_type_lang_key, self::LANG_MODULE), self::dic()->ctrl()
-                ->getLinkTarget($this, self::CMD_LIST_REQUESTS));
+            self::dic()->tabs()->{empty($this->obj_ref_id) ? "addTab" : "addSubTab"}(self::TAB_REQUESTS . $requests_type,
+                self::plugin()->translate("type_" . $requests_type_lang_key, self::LANG_MODULE), self::dic()->ctrl()
+                    ->getLinkTarget($this, self::CMD_LIST_REQUESTS));
         }
         self::dic()->ctrl()->setParameter($this, self::GET_PARAM_REQUESTS_TYPE, $this->requests_type);
     }
@@ -192,8 +195,10 @@ class RequestsGUI
      */
     protected function listRequests()/*: void*/
     {
-        self::dic()->tabs()->activateTab(self::TAB_REQUESTS . self::REQUESTS_TYPE_ALL);
-        self::dic()->tabs()->activateSubTab(self::TAB_REQUESTS . $this->requests_type);
+        if (!empty($this->obj_ref_id)) {
+            self::dic()->tabs()->activateTab(self::TAB_REQUESTS . self::REQUESTS_TYPE_ALL);
+        }
+        self::dic()->tabs()->{empty($this->obj_ref_id) ? "activateTab" : "activateSubTab"}(self::TAB_REQUESTS . $this->requests_type);
 
         $table = self::srUserEnrolment()->enrolmentWorkflow()->requests()->factory()->newTableInstance($this);